+++ example/android/app/src/main/kotlin/com/cielo_flutter/example/cielo_flutter_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.cielo_flutter.example.cielo_flutter_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}