+++ lib/src/apps/sop/cielo_sop_api.dart
@@ -3,16 +3,15 @@ import 'dart:convert';
 import 'package:cielo_flutter/cielo_flutter.dart';
 import 'package:http/http.dart' as http;
 
-
+/// The API service for Silent Order Post product.
 class CieloSOPApi {
-
+  /// The options used in the core's sdk.
   final CieloOptions options;
+
+  /// The options used for Silent Order Post sdk.
   final CieloSOPOptions sopOptions;
 
-  const CieloSOPApi({
-    required this.options,
-    required this.sopOptions
-  });
+  const CieloSOPApi({required this.options, required this.sopOptions});
 
   /// Returns the base URL of Cielo API for the selected provider and environment.
   String get _baseUrl {
@@ -44,7 +43,6 @@ class CieloSOPApi {
     required CieloCard card,
     required CieloSOPOptions options,
   }) async {
-
     final apiUrl = '$_baseUrl/post/api/public/v1/card';
     var headers = {
       'Content-type': 'application/x-www-form-urlencoded',
@@ -63,14 +61,14 @@ class CieloSOPApi {
       return '$key=$value';
     }).join('&');
 
-    var response = await http.post(Uri.parse(apiUrl), headers: headers, body: r);
+    var response =
+        await http.post(Uri.parse(apiUrl), headers: headers, body: r);
 
     if (response.statusCode == 201) {
       return jsonDecode(response.body);
     } else {
-      throw CieloAPIException(code: response.statusCode, message: response.reasonPhrase);
+      throw CieloAPIException(
+          code: response.statusCode, message: response.reasonPhrase);
     }
-
   }
-
 }