+++ lib/src/cielo.dart
@@ -2,7 +2,6 @@ import 'package:cielo_flutter/cielo_flutter.dart';
 import 'package:cielo_flutter/src/apps/sop/cielo_sop.dart';
 
 class Cielo {
-
   Cielo._();
   static Cielo _instance = Cielo._();
 
@@ -64,8 +63,8 @@ class Cielo {
     return _instance._sop!;
   }
 
+  /// Reset Cielo SDK from all configurations.
   static void dispose() {
     _instance = Cielo._();
   }
-
 }