+++ test/src/util/luhn_test.dart
@@ -2,7 +2,6 @@ import 'package:cielo_flutter/src/core/utils/luhn.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
-
   test('accepts a valid card number', () {
     const validCard = '4111111111111111';
     expect(validateCardNumberMod10(validCard), true);
@@ -12,5 +11,4 @@ void main() {
     const invalidCard = '4111111111111211';
     expect(validateCardNumberMod10(invalidCard), false);
   });
-
 }